add prevent-umount-on-pause
If vdr pauses replaying a file it closes all filehandles, which allows autofs to umount a nfs share although it might be needed later, when the user decides to continue replaying a recording. This script opens a file in the directory of the recording and keeps it open until replay ist stopped (or vdr exits).